requirements.txt
click==7.1.2
Flask==1.1.2
Flask-MySQL==1.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyMySQL==0.10.0
PyYAML==5.3.1
Werkzeug==1.0.1
